Show a summary of the function/macro or variable
This solves a number of problems:

1. It gives a link to the source code at the top of the buffer, for
users who want to jump to that quickly. See discussion in #67.

2. It favours giving the filename of the source code in the beginning
of the buffer, as discussed in #69. However, we still show the full
path in the source code section. This avoids confusion when users have
upgraded built-in libraries (e.g. org-mode in ~/.emacs.d/elpa).

3. It ensures that buttons to source code files always jump to the
definition location wherever possible.

4. It makes finding the source code more robust if the C source does
not match the current Emacs version.

5. It highlights functions that are interactive or autoloaded, and
links to the relevant parts of the manual to clarify what that means.